requirements.txt
pandas
numpy
matplotlib
torch
pillow
torchvision
pytest
doit
prospector
black
scikit-learn
opencv-python
tqdm